+++ examples/blueprints2-example/crude_blueprint.h
@@ -405,311 +405,7 @@ constexpr inline uint32_t fnv_1a_hash(const char (&string)[N])
         return GetStaticTypeInfo(); \
     }
 
-
-
-//
-// -------[ Generic Node Types ]-------
-//
-
-struct ConstBoolNode final : Node
-{
-    CRUDE_BP_NODE(ConstBoolNode)
-
-    ConstBoolNode(Blueprint& blueprint): Node(blueprint, "Const Bool") {}
-
-    span<Pin*> GetOutputPins() override { return m_OutputPins; }
-
-    BoolPin m_Bool = { this };
-
-    Pin* m_OutputPins[1] = { &m_Bool };
-};
-
-struct ConstInt32Node final : Node
-{
-    CRUDE_BP_NODE(ConstInt32Node)
-
-    ConstInt32Node(Blueprint& blueprint): Node(blueprint, "Const Int32") {}
-
-    span<Pin*> GetOutputPins() override { return m_OutputPins; }
-
-    Int32Pin m_Int32 = { this };
-
-    Pin* m_OutputPins[1] = { &m_Int32 };
-};
-
-struct ConstFloatNode final : Node
-{
-    CRUDE_BP_NODE(ConstFloatNode)
-
-    ConstFloatNode(Blueprint& blueprint): Node(blueprint, "Const Float") {}
-
-    span<Pin*> GetOutputPins() override { return m_OutputPins; }
-
-    FloatPin m_Float = { this };
-
-    Pin* m_OutputPins[1] = { &m_Float };
-};
-
-struct ConstStringNode final : Node
-{
-    CRUDE_BP_NODE(ConstStringNode)
-
-    ConstStringNode(Blueprint& blueprint): Node(blueprint, "Const String") {}
-
-    span<Pin*> GetOutputPins() override { return m_OutputPins; }
-
-    StringPin m_String = { this };
-
-    Pin* m_OutputPins[1] = { &m_String };
-};
-
-struct BranchNode final : Node
-{
-    CRUDE_BP_NODE(BranchNode)
-
-    BranchNode(Blueprint& blueprint): Node(blueprint, "Branch") {}
-
-    FlowPin Execute(Context& context, FlowPin& entryPoint) override
-    {
-        auto value = context.GetPinValue<bool>(m_Condition);
-        if (value)
-            return m_True;
-        else
-            return m_False;
-    }
-
-    span<Pin*> GetInputPins() override { return m_InputPins; }
-    span<Pin*> GetOutputPins() override { return m_OutputPins; }
-
-    FlowPin m_Enter     = { this };
-    BoolPin m_Condition = { this, "Condition" };
-    FlowPin m_True      = { this, "True" };
-    FlowPin m_False     = { this, "False" };
-
-    Pin* m_InputPins[2] = { &m_Enter, &m_Condition };
-    Pin* m_OutputPins[2] = { &m_True, &m_False };
-};
-
-struct DoNNode final : Node
-{
-    CRUDE_BP_NODE(DoNNode)
-
-    DoNNode(Blueprint& blueprint): Node(blueprint, "Do N") {}
-
-    void Reset(Context& context) override
-    {
-        context.SetPinValue(m_Counter, 0);
-    }
-
-    FlowPin Execute(Context& context, FlowPin& entryPoint) override
-    {
-        if (entryPoint.m_Id == m_Reset.m_Id)
-        {
-            context.SetPinValue(m_Counter, 0);
-            return {};
-        }
-
-        auto n = context.GetPinValue<int32_t>(m_N);
-        auto c = context.GetPinValue<int32_t>(m_Counter);
-        if (c >= n)
-            return {};
-
-        context.SetPinValue(m_Counter, c + 1);
-
-        context.PushReturnPoint(entryPoint);
-
-        return m_Exit;
-    }
-
-    span<Pin*> GetInputPins() override { return m_InputPins; }
-    span<Pin*> GetOutputPins() override { return m_OutputPins; }
-
-    FlowPin  m_Enter   = { this, "Enter" };
-    FlowPin  m_Reset   = { this, "Reset" };
-    Int32Pin m_N       = { this, "N" };
-    FlowPin  m_Exit    = { this, "Exit" };
-    Int32Pin m_Counter = { this, "Counter" };
-
-    Pin* m_InputPins[3] = { &m_Enter, &m_N, &m_Reset };
-    Pin* m_OutputPins[2] = { &m_Exit, &m_Counter };
-};
-
-struct FlipFlopNode final : Node
-{
-    CRUDE_BP_NODE(FlipFlopNode)
-
-    FlipFlopNode(Blueprint& blueprint): Node(blueprint, "Flip Flop") {}
-
-    void Reset(Context& context) override
-    {
-        context.SetPinValue(m_IsA, false);
-    }
-
-    FlowPin Execute(Context& context, FlowPin& entryPoint) override
-    {
-        auto isA = !context.GetPinValue<bool>(m_IsA);
-        context.SetPinValue(m_IsA, isA);
-        if (isA)
-            return m_A;
-        else
-            return m_B;
-    }
-
-    span<Pin*> GetInputPins() override { return m_InputPins; }
-    span<Pin*> GetOutputPins() override { return m_OutputPins; }
-
-    FlowPin m_Enter = { this, "Enter" };
-    FlowPin m_A     = { this, "A" };
-    FlowPin m_B     = { this, "B" };
-    BoolPin m_IsA   = { this, "Is A" };
-
-    Pin* m_InputPins[1] = { &m_Enter };
-    Pin* m_OutputPins[3] = { &m_A, &m_B, &m_IsA };
-};
-
-struct ToStringNode final : Node
-{
-    CRUDE_BP_NODE(ToStringNode)
-
-    ToStringNode(Blueprint& blueprint): Node(blueprint, "To String") {}
-
-    FlowPin Execute(Context& context, FlowPin& entryPoint) override
-    {
-        string result;
-        switch (m_Value.GetType())
-        {
-            case PinType::Any:    break;
-            case PinType::Flow:   break;
-            case PinType::Bool:   result = context.GetPinValue<bool>(m_Value) ? "true" : "false"; break;
-            case PinType::Int32:  result = std::to_string(context.GetPinValue<int32_t>(m_Value)); break;
-            case PinType::Float:  result = std::to_string(context.GetPinValue<float>(m_Value)); break;
-            case PinType::String: result = context.GetPinValue<string>(m_Value); break;
-        }
-
-        context.SetPinValue(m_String, std::move(result));
-
-        return m_Exit;
-    }
-
-    span<Pin*> GetInputPins() override { return m_InputPins; }
-    span<Pin*> GetOutputPins() override { return m_OutputPins; }
-
-    FlowPin   m_Enter  = { this };
-    FlowPin   m_Exit   = { this };
-    AnyPin    m_Value  = { this, "Value" };
-    StringPin m_String = { this, string_view("String") };
-
-    Pin* m_InputPins[2] = { &m_Enter, &m_Value };
-    Pin* m_OutputPins[2] = { &m_Exit, &m_String };
-};
-
-# ifdef _WIN32
-extern "C" __declspec(dllimport) void __stdcall OutputDebugStringA(const char* string);
-# else
-#     include <stdio.h>
-# endif
-
-struct PrintNode final : Node
-{
-    CRUDE_BP_NODE(PrintNode)
-
-    PrintNode(Blueprint& blueprint): Node(blueprint, "Print") {}
-
-    FlowPin Execute(Context& context, FlowPin& entryPoint) override
-    {
-        auto value = context.GetPinValue<string>(m_String);
-
-# ifdef _WIN32
-        OutputDebugStringA("PrintNode: ");
-        OutputDebugStringA(value.c_str());
-        OutputDebugStringA("\n");
-# else
-        printf("PrintNode: %s\n", value.c_str());
-# endif
-
-        return m_Exit;
-    }
-
-    span<Pin*> GetInputPins() override { return m_InputPins; }
-    span<Pin*> GetOutputPins() override { return m_OutputPins; }
-
-    FlowPin   m_Enter  = { this };
-    FlowPin   m_Exit   = { this };
-    StringPin m_String = { this };
-
-    Pin* m_InputPins[2] = { &m_Enter, &m_String };
-    Pin* m_OutputPins[1] = { &m_Exit };
-};
-
-struct EntryPointNode final : Node
-{
-    CRUDE_BP_NODE(EntryPointNode)
-
-    EntryPointNode(Blueprint& blueprint): Node(blueprint, "Start") {}
-
-    FlowPin Execute(Context& context, FlowPin& entryPoint) override
-    {
-        return m_Exit;
-    }
-
-    span<Pin*> GetOutputPins() override { return m_OutputPins; }
-
-    FlowPin m_Exit = { this };
-
-    Pin* m_OutputPins[1] = { &m_Exit };
-};
-
-struct AddNode final : Node
-{
-    CRUDE_BP_NODE(AddNode)
-
-    AddNode(Blueprint& blueprint): Node(blueprint, "Add (const)")
-    {
-        SetType(PinType::Any);
-    }
-
-    PinValue EvaluatePin(const Context& context, const Pin& pin) const override
-    {
-        if (pin.m_Id == m_Result->m_Id)
-        {
-            auto aValue = context.GetPinValue(*m_A);
-            auto bValue = context.GetPinValue(*m_B);
-
-            if (static_cast<PinType>(aValue.index()) != m_Type ||
-                static_cast<PinType>(bValue.index()) != m_Type)
-            {
-                return {}; // Error: Node values must be of same type
-            }
-
-            switch (m_Type)
-            {
-                case PinType::Int32:
-                    return get<int32_t>(aValue) + get<int32_t>(bValue);
-                case PinType::Float:
-                    return get<float>(aValue) + get<float>(bValue);
-                case PinType::String:
-                    return get<float>(aValue) + get<float>(bValue);
-            }
-
-            return {}; // Error: Unsupported type
-        }
-        else
-            return Node::EvaluatePin(context, pin);
-    }
-
-    void SetType(PinType type);
-
-    span<Pin*> GetInputPins() override { return m_InputPins; }
-    span<Pin*> GetOutputPins() override { return m_OutputPins; }
-
-    PinType m_Type = PinType::Any;
-    unique_ptr<Pin> m_A;
-    unique_ptr<Pin> m_B;
-    unique_ptr<Pin> m_Result;
-
-    Pin* m_InputPins[2];
-    Pin* m_OutputPins[1];
-};
+// Look to crude_blueprint_library.h for concrete node examples
 
 
 
@@ -732,20 +428,7 @@ struct NodeRegistry
 private:
     void RebuildTypes();
 
-    NodeTypeInfo m_BuildInNodes[11] =
-    {
-        ConstBoolNode::GetStaticTypeInfo(),
-        ConstInt32Node::GetStaticTypeInfo(),
-        ConstFloatNode::GetStaticTypeInfo(),
-        ConstStringNode::GetStaticTypeInfo(),
-        BranchNode::GetStaticTypeInfo(),
-        DoNNode::GetStaticTypeInfo(),
-        FlipFlopNode::GetStaticTypeInfo(),
-        ToStringNode::GetStaticTypeInfo(),
-        PrintNode::GetStaticTypeInfo(),
-        EntryPointNode::GetStaticTypeInfo(),
-        AddNode::GetStaticTypeInfo(),
-    };
+    vector<NodeTypeInfo> m_BuildInNodes;
 
     vector<NodeTypeInfo> m_CustomNodes;
 
@@ -757,6 +440,8 @@ private:
 //
 // -------[ Blueprint Object ]-------
 //
+
+struct EntryPointNode;
 
 struct Blueprint
 {
